requirements.txt
Django==1.10
IPy==0.83
MySQL-python==1.2.5
PyMySQL==0.7.6
Pygments==2.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.7.0
chardet==3.0.4
cryptography==1.4
decorator==4.1.2
docopt==0.6.2
dwebsocket==0.4.2
enum34==1.1.6
idna==2.6
ipaddress==1.0.16
ipython==5.5.0
ipython-genutils==0.2.0
jenkinsapi==0.3.4
multi-key-dict==2.0.3
paramiko==2.0.2
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.3.0
pickleshare==0.7.4
pipreqs==0.4.9
prompt-toolkit==1.0.15
psutil==5.2.2
ptyprocess==0.5.2
pyasn1==0.1.9
pycparser==2.14
python-jenkins==0.4.15
pytz==2017.2
redis==2.10.5
requests==2.18.4
scandir==1.6
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
web.py==0.21
wsgiref==0.1.2
xlwt==1.1.2
yarg==0.1.9
